migrations/models/3_20230930223809_invoice_model_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "invoice" RENAME COLUMN "customer" TO "customer_email";"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "invoice" RENAME COLUMN "customer_email" TO "customer";"""
