requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==4.0.0
bcrypt==4.0.1
black==23.7.0
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
iso8601==2.0.0
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydantic-settings==2.0.3
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.31.1
tomli==2.0.1
tortoise-orm==0.20.0
typing_extensions==4.7.1
uvicorn==0.23.2
